+++ crates/holochain_p2p/src/spawn/actor.rs
@@ -1019,60 +1019,62 @@ impl HolochainP2pActor {
                         let db_getter = db_getter.clone();
                         let kitsune2 = kitsune2.clone();
                         async move {
-                            let space = kitsune2.clone().space(space_id.clone(), None).await?;
-                            let peer_store = space.peer_store().clone();
-                            let db = db_getter(DnaHash::from_k2_space(&space_id)).await?;
-                            // Prune any expired entries.
-                            db.write_async(|txn| -> DatabaseResult<()> {
-                                let prune_count = txn.execute(sql_peer_meta_store::PRUNE, [])?;
-                                tracing::debug!("Pruned {prune_count} expired rows from peer meta store");
-                                Ok(())
-                            })
-                            .await
-                            .map_err(HolochainP2pError::other)?;
-
-                            // Get agent infos from peer store and compare if there are any up-to-date ones with
-                            // any of the unresponsive URLs.
-                            // That would indicate that the URL was unresponsive temporarily and has become
-                            // responsive again.
-                            let agents = peer_store.get_all().await?;
-                            let urls_to_prune = db
-                                .read_async(move |txn| -> DatabaseResult<Vec<Value>> {
-                                    let mut stmt = txn.prepare(sql_peer_meta_store::GET_ALL_BY_KEY)?;
-                                    let mut rows = stmt.query(
-                                    named_params! {":meta_key":format!("{KEY_PREFIX_ROOT}:{META_KEY_UNRESPONSIVE}")},
-                                    )?;
-                                    let mut urls = Vec::new();
-                                    while let Some(row) = rows.next()? {
-                                        // Expecting is safe here, because the inserted values must have been URLs.
-                                        let peer_url = Url::from_str(row.get::<_, String>(0)?).map_err(|err| DatabaseError::Other(err.into()))?;
-                                        let meta_value = row.get::<_, BytesSql>("meta_value")?;
-                                        let timestamp: kitsune2_api::Timestamp = serde_json::from_slice(&meta_value.0).map_err(|err| DatabaseError::Other(err.into()))?;
-                                        if let Some(agent) = agents
-                                            .iter()
-                                            .find(|agent| agent.url == Some(peer_url.clone()))
-                                        {
-                                            if agent.created_at > timestamp {
-                                                urls.push(Value::Text(peer_url.to_string()));
-                                            }
-                                        }
-                                    }
-                                    Ok(urls)
+                            if let Some(space) = kitsune2.clone().space_if_exists(space_id.clone()).await {
+                                let peer_store = space.peer_store().clone();
+                                let db = db_getter(DnaHash::from_k2_space(&space_id)).await?;
+                                // Prune any expired entries.
+                                db.write_async(|txn| -> DatabaseResult<()> {
+                                    let prune_count = txn.execute(sql_peer_meta_store::PRUNE, [])?;
+                                    tracing::debug!("Pruned {prune_count} expired rows from peer meta store");
+                                    Ok(())
                                 })
                                 .await
                                 .map_err(HolochainP2pError::other)?;
 
-                            // Delete all urls to be pruned from the table.
-                            db.write_async(|txn| -> DatabaseResult<()> {
-                                let values = Rc::new(urls_to_prune);
-                                let mut stmt = txn.prepare(sql_peer_meta_store::DELETE_URLS)?;
-                                stmt.execute(named_params!{":urls": values, ":meta_key": format!("{KEY_PREFIX_ROOT}:{META_KEY_UNRESPONSIVE}")})?;
-                                tracing::debug!("Pruned {} unexpired {KEY_PREFIX_ROOT}:{META_KEY_UNRESPONSIVE} rows from peer meta store because we have newer agent info", values.len());
-                                Ok(())
-                            })
-                            .await
-                            .map_err(HolochainP2pError::other)?;
+                                // Get agent infos from peer store and compare if there are any up-to-date ones with
+                                // any of the unresponsive URLs.
+                                // That would indicate that the URL was unresponsive temporarily and has become
+                                // responsive again.
+                                let agents = peer_store.get_all().await?;
+                                let urls_to_prune = db
+                                    .read_async(move |txn| -> DatabaseResult<Vec<Value>> {
+                                        let mut stmt = txn.prepare(sql_peer_meta_store::GET_ALL_BY_KEY)?;
+                                        let mut rows = stmt.query(
+                                        named_params! {":meta_key":format!("{KEY_PREFIX_ROOT}:{META_KEY_UNRESPONSIVE}")},
+                                        )?;
+                                        let mut urls = Vec::new();
+                                        while let Some(row) = rows.next()? {
+                                            // Expecting is safe here, because the inserted values must have been URLs.
+                                            let peer_url = Url::from_str(row.get::<_, String>(0)?).map_err(|err| DatabaseError::Other(err.into()))?;
+                                            let meta_value = row.get::<_, BytesSql>("meta_value")?;
+                                            let timestamp: kitsune2_api::Timestamp = serde_json::from_slice(&meta_value.0).map_err(|err| DatabaseError::Other(err.into()))?;
+                                            if let Some(agent) = agents
+                                                .iter()
+                                                .find(|agent| agent.url == Some(peer_url.clone()))
+                                            {
+                                                if agent.created_at > timestamp {
+                                                    urls.push(Value::Text(peer_url.to_string()));
+                                                }
+                                            }
+                                        }
+                                        Ok(urls)
+                                    })
+                                    .await
+                                    .map_err(HolochainP2pError::other)?;
 
+                                // Delete all urls to be pruned from the table.
+                                db.write_async(|txn| -> DatabaseResult<()> {
+                                    let values = Rc::new(urls_to_prune);
+                                    let mut stmt = txn.prepare(sql_peer_meta_store::DELETE_URLS)?;
+                                    stmt.execute(named_params!{":urls": values, ":meta_key": format!("{KEY_PREFIX_ROOT}:{META_KEY_UNRESPONSIVE}")})?;
+                                    tracing::debug!("Pruned {} unexpired {KEY_PREFIX_ROOT}:{META_KEY_UNRESPONSIVE} rows from peer meta store because we have newer agent info", values.len());
+                                    Ok(())
+                                })
+                                .await
+                                .map_err(HolochainP2pError::other)?;
+                            } else {
+                                tracing::warn!("Cannot prune expired URLs from peer meta store for k2 space that does not exist with space id {space_id}");
+                            }
                             Ok::<_, HolochainP2pError>(())
                         }
                     });
@@ -1263,8 +1265,9 @@ impl HolochainP2pActor {
         ops: Vec<StoredOp>,
     ) -> HolochainP2pResult<()> {
         self.kitsune
-            .space(space_id, None)
-            .await?
+            .space_if_exists(space_id.clone())
+            .await
+            .ok_or(HolochainP2pError::K2SpaceNotFound(space_id))?
             .inform_ops_stored(ops)
             .await
             .map_err(HolochainP2pError::K2Error)
@@ -1384,10 +1387,12 @@ impl actor::HcP2p for HolochainP2pActor {
 
     fn peer_store(&self, dna_hash: DnaHash) -> BoxFut<'_, HolochainP2pResult<DynPeerStore>> {
         Box::pin(async move {
+            let space_id = dna_hash.to_k2_space();
             Ok(self
                 .kitsune
-                .space(dna_hash.to_k2_space(), None)
-                .await?
+                .space_if_exists(space_id.clone())
+                .await
+                .ok_or(HolochainP2pError::K2SpaceNotFound(space_id))?
                 .peer_store()
                 .clone())
         })
@@ -1426,6 +1431,11 @@ impl actor::HcP2p for HolochainP2pActor {
                 Some(overrides) => self.space_config_override(overrides)?,
                 None => None,
             };
+
+            // Create k2 space with config override.
+            //
+            // This is the only time the space is created,
+            // all other calls to get a space should use `space_if_exists`.
             let space = self
                 .kitsune
                 .space(dna_hash.to_k2_space(), config_override)
@@ -1450,21 +1460,23 @@ impl actor::HcP2p for HolochainP2pActor {
     ) -> BoxFut<'_, HolochainP2pResult<()>> {
         Box::pin(async move {
             let space_id = dna_hash.to_k2_space();
-            let space = self.kitsune.space(space_id.clone(), None).await?;
+            if let Some(space) = self.kitsune.space_if_exists(space_id.clone()).await {
+                space.local_agent_leave(agent_pub_key.to_k2_agent()).await;
 
-            space.local_agent_leave(agent_pub_key.to_k2_agent()).await;
-
-            // If there are no more local agents in this space, then the space can be removed.
-            if space
-                .local_agent_store()
-                .get_all()
-                .await
-                .is_ok_and(|agents| agents.is_empty())
-            {
-                drop(space);
-                if let Err(err) = self.kitsune.remove_space(space_id).await {
-                    tracing::warn!(?err, "Failed to remove space after last agent left");
+                // If there are no more local agents in this space, then the space can be removed.
+                if space
+                    .local_agent_store()
+                    .get_all()
+                    .await
+                    .is_ok_and(|agents| agents.is_empty())
+                {
+                    drop(space);
+                    if let Err(err) = self.kitsune.remove_space(space_id).await {
+                        tracing::warn!(?err, "Failed to remove space after last agent left");
+                    }
                 }
+            } else {
+                tracing::warn!("Cannot leave space that does not exist with space id {space_id}.");
             }
 
             Ok(())
@@ -1488,7 +1500,11 @@ impl actor::HcP2p for HolochainP2pActor {
     ) -> BoxFut<'_, HolochainP2pResult<SerializedBytes>> {
         Box::pin(async move {
             let space_id = dna_hash.to_k2_space();
-            let space = self.kitsune.space(space_id.clone(), None).await?;
+            let space = self
+                .kitsune
+                .space_if_exists(space_id.clone())
+                .await
+                .ok_or(HolochainP2pError::K2SpaceNotFound(space_id))?;
 
             let byte_count = zome_call_params_serialized.0.len();
 
@@ -1535,7 +1551,11 @@ impl actor::HcP2p for HolochainP2pActor {
     ) -> BoxFut<'_, HolochainP2pResult<()>> {
         Box::pin(async move {
             let space_id = dna_hash.to_k2_space();
-            let space = self.kitsune.space(space_id.clone(), None).await?;
+            let space = self
+                .kitsune
+                .space_if_exists(space_id.clone())
+                .await
+                .ok_or(HolochainP2pError::K2SpaceNotFound(space_id.clone()))?;
 
             let byte_count: usize = target_payload_list.iter().map(|(_, p, _)| p.0.len()).sum();
 
@@ -1605,9 +1625,13 @@ impl actor::HcP2p for HolochainP2pActor {
                     .await?;
             }
 
-            let space = dna_hash.to_k2_space();
+            let space_id = dna_hash.to_k2_space();
 
-            let space = self.kitsune.space(space, None).await?;
+            let space = self
+                .kitsune
+                .space_if_exists(space_id.clone())
+                .await
+                .ok_or(HolochainP2pError::K2SpaceNotFound(space_id))?;
 
             // -- actually publish the op hashes -- //
 
@@ -1652,7 +1676,11 @@ impl actor::HcP2p for HolochainP2pActor {
     ) -> BoxFut<'_, HolochainP2pResult<()>> {
         Box::pin(async move {
             let space_id = dna_hash.to_k2_space();
-            let space = self.kitsune.space(space_id.clone(), None).await?;
+            let space = self
+                .kitsune
+                .space_if_exists(space_id.clone())
+                .await
+                .ok_or(HolochainP2pError::K2SpaceNotFound(space_id))?;
 
             let peers = self
                 .get_peers_for_location(&space, basis_hash.get_loc())
@@ -1703,7 +1731,12 @@ impl actor::HcP2p for HolochainP2pActor {
     ) -> BoxFut<'_, HolochainP2pResult<Vec<WireOps>>> {
         Box::pin(async move {
             let space_id = dna_hash.to_k2_space();
-            let space = self.kitsune.space(space_id.clone(), None).await?;
+            let space = self
+                .kitsune
+                .space_if_exists(space_id.clone())
+                .await
+                .ok_or(HolochainP2pError::K2SpaceNotFound(space_id.clone()))?;
+
             let loc = dht_hash.get_loc();
             let agents = self
                 .get_random_peers_for_location("get", &space, loc, &options)
@@ -1777,7 +1810,11 @@ impl actor::HcP2p for HolochainP2pActor {
     ) -> BoxFut<'_, HolochainP2pResult<Vec<WireLinkOps>>> {
         Box::pin(async move {
             let space_id = dna_hash.to_k2_space();
-            let space = self.kitsune.space(space_id.clone(), None).await?;
+            let space = self
+                .kitsune
+                .space_if_exists(space_id.clone())
+                .await
+                .ok_or(HolochainP2pError::K2SpaceNotFound(space_id))?;
             let loc = link_key.base.get_loc();
             let agents = self
                 .get_random_peers_for_location(
@@ -1839,7 +1876,11 @@ impl actor::HcP2p for HolochainP2pActor {
     ) -> BoxFut<'_, HolochainP2pResult<CountLinksResponse>> {
         Box::pin(async move {
             let space_id = dna_hash.to_k2_space();
-            let space = self.kitsune.space(space_id.clone(), None).await?;
+            let space = self
+                .kitsune
+                .space_if_exists(space_id.clone())
+                .await
+                .ok_or(HolochainP2pError::K2SpaceNotFound(space_id))?;
             let loc = query.base.get_loc();
             let agents = self
                 .get_random_peers_for_location("count_links", &space, loc, &options)
@@ -1891,7 +1932,11 @@ impl actor::HcP2p for HolochainP2pActor {
     ) -> BoxFut<'_, HolochainP2pResult<Vec<AgentActivityResponse>>> {
         Box::pin(async move {
             let space_id = dna_hash.to_k2_space();
-            let space = self.kitsune.space(space_id.clone(), None).await?;
+            let space = self
+                .kitsune
+                .space_if_exists(space_id.clone())
+                .await
+                .ok_or(HolochainP2pError::K2SpaceNotFound(space_id))?;
             let loc = agent.get_loc();
             let agents = self
                 .get_random_peers_for_location(
@@ -1965,7 +2010,11 @@ impl actor::HcP2p for HolochainP2pActor {
     ) -> BoxFut<'_, HolochainP2pResult<Vec<MustGetAgentActivityResponse>>> {
         Box::pin(async move {
             let space_id = dna_hash.to_k2_space();
-            let space = self.kitsune.space(space_id.clone(), None).await?;
+            let space = self
+                .kitsune
+                .space_if_exists(space_id.clone())
+                .await
+                .ok_or(HolochainP2pError::K2SpaceNotFound(space_id))?;
             let loc = author.get_loc();
             let agents = self
                 .get_random_peers_for_location("must_get_agent_activity", &space, loc, &options)
@@ -2021,7 +2070,11 @@ impl actor::HcP2p for HolochainP2pActor {
     ) -> BoxFut<'_, HolochainP2pResult<()>> {
         Box::pin(async move {
             let space_id = dna_hash.to_k2_space();
-            let space = self.kitsune.space(space_id.clone(), None).await?;
+            let space = self
+                .kitsune
+                .space_if_exists(space_id.clone())
+                .await
+                .ok_or(HolochainP2pError::K2SpaceNotFound(space_id))?;
 
             let agent_id = to_agent.to_k2_agent();
 
@@ -2083,7 +2136,11 @@ impl actor::HcP2p for HolochainP2pActor {
         Box::pin(async move {
             let loc = basis.get_loc();
             let space_id = dna_hash.to_k2_space();
-            let space = self.kitsune.space(space_id.clone(), None).await?;
+            let space = self
+                .kitsune
+                .space_if_exists(space_id.clone())
+                .await
+                .ok_or(HolochainP2pError::K2SpaceNotFound(space_id))?;
 
             for agent in space.local_agent_store().get_all().await? {
                 if agent.get_cur_storage_arc().contains(loc) {
@@ -2103,7 +2160,11 @@ impl actor::HcP2p for HolochainP2pActor {
     ) -> BoxFut<'_, HolochainP2pResult<()>> {
         Box::pin(async move {
             let space_id = dna_hash.to_k2_space();
-            let space = self.kitsune.space(space_id.clone(), None).await?;
+            let space = self
+                .kitsune
+                .space_if_exists(space_id.clone())
+                .await
+                .ok_or(HolochainP2pError::K2SpaceNotFound(space_id.clone()))?;
 
             let mut peer_urls = Vec::with_capacity(agents.len());
             for agent in agents {
@@ -2193,7 +2254,11 @@ impl actor::HcP2p for HolochainP2pActor {
                     let all_space_ids = self.kitsune.list_spaces();
                     let mut spaces = Vec::with_capacity(all_space_ids.len());
                     for space_id in all_space_ids {
-                        spaces.push((space_id.clone(), self.kitsune.space(space_id, None).await?));
+                        if let Some(space) = self.kitsune.space_if_exists(space_id.clone()).await {
+                            spaces.push((space_id.clone(), space));
+                        } else {
+                            tracing::warn!("Cannot dump network metrics for space that does not exist with space id {space_id}.");
+                        }
                     }
 
                     spaces
@@ -2250,7 +2315,11 @@ impl actor::HcP2p for HolochainP2pActor {
     ) -> BoxFut<'_, HolochainP2pResult<Vec<kitsune2_api::DhtArc>>> {
         Box::pin(async move {
             let space_id = dna_hash.to_k2_space();
-            let space = self.kitsune.space(space_id, None).await?;
+            let space = self
+                .kitsune
+                .space_if_exists(space_id.clone())
+                .await
+                .ok_or(HolochainP2pError::K2SpaceNotFound(space_id))?;
 
             Ok(space
                 .local_agent_store()